chore(tests) Moving moveTemplate tests to a separate workflow
