+++ editor/package.json
@@ -17,8 +17,10 @@
     "lint": "npx eslint \"src/**/*.{js,ts,tsx}\" --quiet",
     "lint-loud": "npx eslint \"src/**/*.{js,ts,tsx}\"",
     "check-dependency-cruiser": "depcruise --config .dependency-cruiser.js src",
-    "test": "jest --config ./jest.config.js",
+    "test": "jest --config ./jest.config.js --testPathIgnorePatterns \"move-template.spec.browser\"",
     "test-ci": "NODE_OPTIONS=--max_old_space_size=6656 xvfb-run -a npm run test -- --maxWorkers=1",
+    "test-move-template-only": "jest --config ./jest.config.js move-template.spec.browser",
+    "test-move-template-only-ci": "NODE_OPTIONS=--max_old_space_size=6656 xvfb-run -a npm run test-move-template-only -- --maxWorkers=1",
     "test-debug": "node --inspect-brk node_modules/.bin/jest --config ./jest.config.js --runInBand",
     "test-watch": "npm run test -- --watch",
     "test-update-snapshot": "npm run test -- --updateSnapshot",
